requirements.txt
django
pillow
django-crispy-forms
pymysql
mysqlclient
django-summernote
django-filter